Make uncheckedSendable a reusable extension
